Extract review actions from restaurant actions into its own file
